+++ app/components/showHidePassword/index.js
@@ -1,0 +1,3 @@
+import ShowHidePassword from './showHidePassword';
+
+export default ShowHidePassword;