+++ CHANGELOG.md
@@ -5,6 +5,11 @@ It is based on Tidepool Blip 1.27.
 ## UNRELEASED
 ### Fixed
 - PT-869 Fix labelling issue in validation errors
+  
+### Added
+- PN-10 Add Show/hide password in Signup page
+- PN-9 Add Show/hide password in Login page
+- PT-412 Add option to display password when changing password
 
 ### Changed
  - PT-836 Display list of patients to clinician as a default behaviour